requirements.txt
googletrans==3.1.0a0
disnake==2.9.0
anyio
wexpect
python-dotenv==1.0.0